Give a user property the value type it says it has
FBX states a user property's type and then its value, and a reader is entitled
to hold them to each other. Two of ours did not.

A boolean track's constant was declared `bool` and written as a float64.
Blender asserts that a `bool` is an int32, and an assertion inside its importer
is not a skipped property or a skipped object -- it abandons the file. Every
mesh with a boolean track failed to open: an emitter switching on, anything that
hides itself, `TestNifFile_Animated_LE` and the eleven fixtures beside it.

And an alpha threshold was declared `Short`, which is not an FBX user property
type at all. Blender warns and drops the value; another reader may do worse.

Neither shows up in a NIF -> FBX -> NIF trip, because this reads back what it
writes whatever it calls it, and the SDK probe opens the file without asking
what a property claims to be. It took putting Blender in the middle to find
them, which is what the test in the next commit does.

+++ src/NIFBX/Fbx/FbxMaterialWriter.cs
@@ -172,7 +172,9 @@ namespace NIFBX.Fbx
                 alpha.EditorAlphaThreshold ? 1 : 0);
 
             // Named for Blender, which surfaces a short here.
-            properties.Set("alpha_test_threshold", "Short", "", FbxProperties.UserFlags, (int)alpha.Threshold);
+            // `int`, not `Short`: FBX has no such user property type, and a reader that
+            // meets one either warns and drops the value or stops reading.
+            properties.Set("alpha_test_threshold", "int", "", FbxProperties.UserFlags, (int)alpha.Threshold);
         }
 
         private static void AddTextures(
